+++ test/conftest.py
@@ -12,26 +12,26 @@ import ufl
 elements = [("Lagrange", ufl.interval, 1),
             ("Lagrange", ufl.triangle, 1),
             ("Lagrange", ufl.tetrahedron, 1),
-            # ("Lagrange", ufl.quadrilateral, 1),
-            # ("Lagrange", ufl.hexahedron, 1),
+            ("Lagrange", ufl.quadrilateral, 1),
+            ("Lagrange", ufl.hexahedron, 1),
             ("Lagrange", ufl.interval, 2),
             ("Lagrange", ufl.triangle, 2),
             ("Lagrange", ufl.tetrahedron, 2),
-            # ("Lagrange", ufl.quadrilateral, 2),
-            # ("Lagrange", ufl.hexahedron, 2),
+            ("Lagrange", ufl.quadrilateral, 2),
+            ("Lagrange", ufl.hexahedron, 2),
             ("Lagrange", ufl.interval, 3),
             ("Lagrange", ufl.triangle, 3),
             ("Lagrange", ufl.tetrahedron, 3),
-            # ("Lagrange", ufl.quadrilateral, 3),
+            ("Lagrange", ufl.quadrilateral, 3),
             # ("Lagrange", ufl.quadrilateral, 3, None, None, "spectral"),
-            # ("Lagrange", ufl.hexahedron, 3),
+            ("Lagrange", ufl.hexahedron, 3),
             # ("Lagrange", ufl.hexahedron, 3, None, None, "spectral"),
-            # ("Brezzi-Douglas-Marini", ufl.triangle, 1),
-            # ("Brezzi-Douglas-Marini", ufl.tetrahedron, 1),
-            # ("Brezzi-Douglas-Marini", ufl.triangle, 2),
-            # ("Brezzi-Douglas-Marini", ufl.tetrahedron, 2),
-            # ("Brezzi-Douglas-Marini", ufl.triangle, 3),
-            # ("Brezzi-Douglas-Marini", ufl.tetrahedron, 3),
+            ("Brezzi-Douglas-Marini", ufl.triangle, 1),
+            ("Brezzi-Douglas-Marini", ufl.tetrahedron, 1),
+            ("Brezzi-Douglas-Marini", ufl.triangle, 2),
+            ("Brezzi-Douglas-Marini", ufl.tetrahedron, 2),
+            ("Brezzi-Douglas-Marini", ufl.triangle, 3),
+            ("Brezzi-Douglas-Marini", ufl.tetrahedron, 3),
             ("Raviart-Thomas", ufl.triangle, 1),
             ("Raviart-Thomas", ufl.tetrahedron, 1),
             ("Raviart-Thomas", ufl.triangle, 2),
